refactor(web): Remove undesired in-source vitest test
