+++ web/src/utils/api.ts
@@ -75,7 +75,7 @@ const hasReportedFailedFetchMessage = (error: unknown): boolean => {
   return REPORTED_FAILED_FETCH_MESSAGE.test(error.message);
 };
 
-const isNetworkConnectivityError = (error: unknown): boolean => {
+export const isNetworkConnectivityError = (error: unknown): boolean => {
   if (!(error instanceof TRPCClientError)) return false;
 
   // tRPC server errors and infrastructure responses have response metadata.
@@ -88,71 +88,6 @@ const isNetworkConnectivityError = (error: unknown): boolean => {
     hasReportedFailedFetchMessage(error)
   );
 };
-
-/* eslint-disable @repo/no-in-source-vitest */
-const vitest = import.meta.vitest;
-
-if (vitest && typeof vitest === "object") {
-  const { describe, expect, it } = vitest;
-
-  describe("isNetworkConnectivityError", () => {
-    it("detects the reported failed fetch error without a response", () => {
-      const error = TRPCClientError.from(new TypeError("Failed to fetch"));
-
-      expect(isNetworkConnectivityError(error)).toBe(true);
-    });
-
-    it("detects the reported failed fetch error with a hostname suffix", () => {
-      const error = TRPCClientError.from(
-        new TypeError("Failed to fetch (cloud.langfuse.com)"),
-      );
-
-      expect(isNetworkConnectivityError(error)).toBe(true);
-    });
-
-    it("does not treat other network failures as connectivity errors", () => {
-      const error = TRPCClientError.from(new TypeError("Load failed"));
-
-      expect(isNetworkConnectivityError(error)).toBe(false);
-    });
-
-    it("does not treat tRPC server errors as connectivity errors", () => {
-      const error = TRPCClientError.from({
-        error: {
-          code: -32603,
-          message: "Internal server error",
-          data: {
-            code: "INTERNAL_SERVER_ERROR",
-            httpStatus: 500,
-            path: "events.all",
-          },
-        },
-      });
-
-      expect(isNetworkConnectivityError(error)).toBe(false);
-    });
-
-    it("does not treat response parsing errors as connectivity errors", () => {
-      const error = TRPCClientError.from(
-        new SyntaxError("Unexpected token <"),
-        {
-          meta: {
-            response: new Response("<html></html>", { status: 502 }),
-          },
-        },
-      );
-
-      expect(isNetworkConnectivityError(error)).toBe(false);
-    });
-
-    it("does not treat non-tRPC errors as connectivity errors", () => {
-      expect(isNetworkConnectivityError(new TypeError("Failed to fetch"))).toBe(
-        false,
-      );
-    });
-  });
-}
-/* eslint-enable @repo/no-in-source-vitest */
 
 /**
  * tRPC serializes query input into the GET URL. For reads whose input scales with
